requirements.txt
click==8.0.3
Flask==2.0.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
pyobjc-core==8.2
pyobjc-framework-Cocoa==8.2
pyobjc-framework-Quartz==8.2
Werkzeug==2.0.2
